rename package name & remove dockers
