--- example/helm/bcos-logger/templates/service.yaml
@@ -1,16 +1,0 @@
-apiVersion: v1
-kind: Service
-metadata:
-  name: {{ include "bcos-logger.fullname" . }}
-  labels:
-    {{- include "bcos-logger.labels" . | nindent 4 }}
-spec:
-  type: {{ .Values.service.type }}
-  ports:
-    - port: {{ .Values.service.port }}
-      targetPort: http
-      nodePort: {{ .Values.service.nodePort }}
-      protocol: TCP
-      name: http
-  selector:
-    {{- include "bcos-logger.selectorLabels" . | nindent 4 }}
