--- example/src/main/java/com/lrhealth/bcos/blockchainlogger/BlockchainLoggerApplication.java
@@ -1,13 +1,0 @@
-package com.lrhealth.bcos.blockchainlogger;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class BlockchainLoggerApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(BlockchainLoggerApplication.class, args);
-	}
-
-}
