--- example/src/main/java/com/lrhealth/bcos/blockchainlogger/controller/Swagger2Config.java
@@ -1,27 +1,0 @@
-package com.lrhealth.bcos.blockchainlogger.controller;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import springfox.documentation.builders.ApiInfoBuilder;
-import springfox.documentation.builders.PathSelectors;
-import springfox.documentation.builders.RequestHandlerSelectors;
-import springfox.documentation.service.ApiInfo;
-import springfox.documentation.spi.DocumentationType;
-import springfox.documentation.spring.web.plugins.Docket;
-import springfox.documentation.swagger2.annotations.EnableSwagger2WebMvc;
-
-@Configuration
-@EnableSwagger2WebMvc
-public class Swagger2Config {
-    @Bean
-    public Docket createRestApi() {
-        return new Docket(DocumentationType.SWAGGER_2).useDefaultResponseMessages(false).apiInfo(apiInfo()).select()
-                .apis(RequestHandlerSelectors.any()).paths(PathSelectors.regex("(?!/error.*).*")).build();
-    }
-
-    private ApiInfo apiInfo() {
-        return new ApiInfoBuilder().title("区块链 - 日志存证系统").version("1.0.0").build();
-    }
-
-}
