+++ example/src/main/java/org/fisco/bcos/blockchainlogger/BlockchainLoggerApplication.java
@@ -1,0 +1,13 @@
+package org.fisco.bcos.blockchainlogger;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BlockchainLoggerApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(BlockchainLoggerApplication.class, args);
+	}
+
+}
