+++ example/src/main/resources/application-local.properties
@@ -26,13 +26,13 @@ bcos.network.peers[1]=127.0.0.1:20201
 
 
 bcos.account.keyStoreDir=account
-bcos.account.accountFilePath=conf/0x3b525049e83aac4ee30cf662a47b64223d6edc4d.pem
+bcos.account.accountFilePath=conf/0x22fec9d7e121960e7972402789868962238d8037.pem
 bcos.account.accountFileFormat=pem
-bcos.account.accountAddress=0x3b525049e83aac4ee30cf662a47b64223d6edc4d
+bcos.account.accountAddress=0x22fec9d7e121960e7972402789868962238d8037
 # account.password=123456
 
 bcos.threadPool.channelProcessorThreadSize=16
 bcos.threadPool.receiptProcessorThreadSize=16
 bcos.threadPool.maxBlockingQueueSize=102400
 
-bcos.contract.logger.address=0xab576c5a33c6fb56a85c8ecec94dadef6afc56e5
+bcos.contract.logger.address=
