+++ example/src/main/resources/conf/0x22fec9d7e121960e7972402789868962238d8037.pem
@@ -1,0 +1,5 @@
+-----BEGIN PRIVATE KEY-----
+MIGNAgEAMBAGByqGSM49AgEGBSuBBAAKBHYwdAIBAQQgtzyrrDDj2q7LRD6iaFqI
+ZaCHsuwjUQS94iok+U6QyP6gBwYFK4EEAAqhRANCAASGZahMzkRWI8T2GYECSa4x
+maBkQG9p7ds4uO7BiFNRUix1hl1ZGnA63lHkOy+Jp6Zrmh4bxIoDsuO2D6RUI/fp
+-----END PRIVATE KEY-----
