+++ example/src/main/resources/conf/ca.crt
@@ -1,12 +1,12 @@
 -----BEGIN CERTIFICATE-----
-MIIBvzCCAWSgAwIBAgIUagAdTk8Ls5W1aPPbxLfxNuQnm7kwCgYIKoZIzj0EAwIw
+MIIBvzCCAWSgAwIBAgIUGGZ8gOCz5qpXbBc4WFERSruIO60wCgYIKoZIzj0EAwIw
 NTEOMAwGA1UEAwwFY2hhaW4xEzARBgNVBAoMCmZpc2NvLWJjb3MxDjAMBgNVBAsM
-BWNoYWluMCAXDTIxMDUwNzAxNTQ0N1oYDzIxMjEwNDEzMDE1NDQ3WjA1MQ4wDAYD
+BWNoYWluMCAXDTIxMDYxMjEzNDUxOFoYDzIxMjEwNTE5MTM0NTE4WjA1MQ4wDAYD
 VQQDDAVjaGFpbjETMBEGA1UECgwKZmlzY28tYmNvczEOMAwGA1UECwwFY2hhaW4w
-VjAQBgcqhkjOPQIBBgUrgQQACgNCAASSIvdZG+W5EvY99cloZ0G9c0RNo2dQcKir
-nzmWfnLhKL2zqKGlL7MfPQ3g+xuGvCWZ4abdRb1N+yc2pN4uedlco1MwUTAdBgNV
-HQ4EFgQUypBefrx9yqPS+SndXKGUtOz3mXswHwYDVR0jBBgwFoAUypBefrx9yqPS
-+SndXKGUtOz3mXswDwYDVR0TAQH/BAUwAwEB/zAKBggqhkjOPQQDAgNJADBGAiEA
-/zo+JEMtBsFNuJIljR6zRtOnmyv7yvf8R6Jcn0SLa2kCIQD85QjLfDqKqP6ob6i2
-QmjUlKOMcFQV7+xHRtbLTRq5nA==
+VjAQBgcqhkjOPQIBBgUrgQQACgNCAAQcabNhRatBxCOIyFnT+SQMERI8g/MVhY+/
+HMkNVryko4s4vsTA5XJezF0R9n828h+cuUE8i3uaAzFdCt3pzOa6o1MwUTAdBgNV
+HQ4EFgQUIHc9nRuClQxs2J6yCZeEJwW4lKEwHwYDVR0jBBgwFoAUIHc9nRuClQxs
+2J6yCZeEJwW4lKEwDwYDVR0TAQH/BAUwAwEB/zAKBggqhkjOPQQDAgNJADBGAiEA
+xbqLMZ78pCEE2xYsiR+rkVoS8yTfltwSLP2ca65Z1NYCIQCP2UDVS0nmCgH+jo0w
+AfS4zhxNcwuLQDaPlj5yo/nf3w==
 -----END CERTIFICATE-----
