+++ example/src/main/resources/conf/node.key
@@ -1,5 +1,5 @@
 -----BEGIN PRIVATE KEY-----
-MIGEAgEAMBAGByqGSM49AgEGBSuBBAAKBG0wawIBAQQgimSQHgvIA/pNKtBqM0Vp
-TddxEx+m87GsAja3Hq/EV1ChRANCAAQBQtxFpn5V6xPfT99sW/xski4RhZbXI7ox
-GmGI328Pj+ULDdXtM6ZyXbJqkUDcLAyMn4tK2XhJYEgzwbyy9Rye
+MIGEAgEAMBAGByqGSM49AgEGBSuBBAAKBG0wawIBAQQg9vvQa4U9YGGaz6/vNXcJ
+cFvHlBdIIzR+cOzw+cbJe/ihRANCAARHMmyyBElRuRog4eM+x/h9TQ7vngy9Aiz1
+gfUrnLSoXTP0S4gMoAYr34Cv9GEKjvT6ll8VDKWVsdqSpwTuPwdy
 -----END PRIVATE KEY-----
