+++ example/src/main/resources/conf/sdk.publickey
@@ -1,1 +1,1 @@
-0142dc45a67e55eb13df4fdf6c5bfc6c922e118596d723ba311a6188df6f0f8fe50b0dd5ed33a6725db26a9140dc2c0c8c9f8b4ad97849604833c1bcb2f51c9e
+47326cb2044951b91a20e1e33ec7f87d4d0eef9e0cbd022cf581f52b9cb4a85d33f44b880ca0062bdf80aff4610a8ef4fa965f150ca595b1da92a704ee3f0772
