+++ example/tools/updateContractJava.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 CUR_PATH=$(pwd);
-./sol2java.sh com.lrhealth.bcos.blockchainlogger.contract $CUR_PATH/../src/main/resources/contract $CUR_PATH/../src/main/java
+./sol2java.sh org.fisco.bcos.blockchainlogger.contract $CUR_PATH/../src/main/resources/contract $CUR_PATH/../src/main/java
